+++ arinc665_qt/media/FileWidget.cpp
@@ -42,11 +42,11 @@ void FileWidget::selectedFile(
   {
     ui->nameLineEdit->setText( QString::fromUtf8(
       fileV->name().data(),
-      fileV->name().length()));
+      static_cast< int >( fileV->name().length() ) ) );
 
     ui->partNumberLineEdit->setText( QString::fromUtf8(
       fileV->partNumber().data(),
-      fileV->partNumber().length()));
+      static_cast< int >( fileV->partNumber().length() ) ) );
   }
 
   switch ( fileV->fileType())
